permisos a las vistas, creacion del archivo permisos
se le escribio la condicion para permisos de usuarios em las vistas, y
la classe que maneja esos permisos, no se escribio en todas las vistas,
unas no estan sincronizadas con la master, y la vista denuncias que
trata con la aplicacion movil no tienen aun permisos

+++ APITX/app/vistas/vistaRecurso.py
@@ -11,41 +11,48 @@ def lista_objetos(request):
     """
     Lista de todas las Buses, o crear una nueva
     """
-    if request.method == 'GET':
-        objeto = TxdRecurso.objects.all()
-        serializador = TxdRecursoS(objeto, many=True)
-        return Response(serializador.data)
+    if request.user.has_perms(permisos.lista_duenios):
+        if request.method == 'GET':
+            objeto = TxdRecurso.objects.all()
+            serializador = TxdRecursoS(objeto, many=True)
+            return Response(serializador.data)
 
-    elif request.method == 'POST':
-        print request.data
-        serializador = TxdRecursoS(data=request.data)
-        print request.data['direccion']
-        if serializador.is_valid():
-            serializador.save()
-            return Response(serializador.data,status=status.HTTP_201_CREATED)
-        else:
-            return Response(serializador.errors,status=status.HTTP_400_BAD_REQUEST)
+        elif request.method == 'POST':
+            print request.data
+            serializador = TxdRecursoS(data=request.data)
+            print request.data['direccion']
+            if serializador.is_valid():
+                serializador.save()
+                return Response(serializador.data,status=status.HTTP_201_CREATED)
+            else:
+                return Response(serializador.errors,status=status.HTTP_400_BAD_REQUEST)
+    else:
+        return Response(status=status.HTTP_403_NOT_FOUND)
+
 @api_view(['GET', 'PUT','DELETE'])
 def detalle_objetos(request, pk):
     """
     Actualiza, elimina un objeto segun su id
     """
-    try:
-        objeto = TxdRecurso.objects.get(pk=pk)
-    except ObjectDoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+    if request.user.has_perms(permisos.lista_duenios):
+        try:
+            objeto = TxdRecurso.objects.get(pk=pk)
+        except ObjectDoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
 
-    if request.method == 'GET':
-        serializador = TxdRecursoS(objeto)
-        return Response(serializador.data)
-
-    elif request.method == 'PUT':
-        serializador = TxdRecursoS(objeto, data=request.data)
-        if serializador.is_valid():
-            serializador.save()
+        if request.method == 'GET':
+            serializador = TxdRecursoS(objeto)
             return Response(serializador.data)
-        return Response(serializador.errors,status=status.HTTP_400_BAD_REQUEST)
 
-    elif request.method == 'DELETE':
-        objeto.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        elif request.method == 'PUT':
+            serializador = TxdRecursoS(objeto, data=request.data)
+            if serializador.is_valid():
+                serializador.save()
+                return Response(serializador.data)
+            return Response(serializador.errors,status=status.HTTP_400_BAD_REQUEST)
+
+        elif request.method == 'DELETE':
+            objeto.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+    else:
+        return Response(status=status.HTTP_403_NOT_FOUND)
